Wiki: Update and expand specific pages (#4)
* wiki: Change default scheduler and wording

EEVDF is the current default scheduler, so it should be reflected in the wiki as well. I have changed the wording a bit in the first paragraph.

* wiki: Update kernel page to latest information

Update the kernel page to latest information reflected in README.md of linux-cachyos repo

* wiki: Add EEDVF to supported kernel list

* wiki: Add bottles description

Add some description about Bottles, install method and a link to their documentation.

* wiki: Wording in cachyos-performance

* wiki: add protonupqt in proton section

Protonup-qt is a great tool to manage Wine-GE and Proton-GE installations.

* wiki: Add clarification regarding runners

Wine-GE is great for non-steam games and preferable over standard wine due to support of fsync and other technologies. Proton-GE/CachyOs/Experimental are all great options for Steam games. Adding the hint to not use Proton for non-Steam games is generally adviced by GE himself.

* wiki: replace BMQ and MuQSS

Replace BMW and MuQSS in file with BORE and PDS

* wiki: Remove Protonup-qt part

As per request by ptr1337

* wiki: Rebase (?)

* wiki: add some missing part

* wiki: revert changes from previous commit