First pieces of re-structurization astro.config.mjs
This is the first step to creating the structure of our Wiki. I decided to make the first part of the wiki targeted to "How to get CachyOS", so it will be easy for users to place CachyOS in their devices. Then I can continue on more complex kinds of stuff like HW, drivers, gaming, etc.