docs: Rename boot loader -> boot manager
When talking about UEFI systems, the correct term to refer to systemd-boot, rEFInd and even GRUB
is boot manager. Boot manager is not the same as a boot loader.

A boot loader handles the task of loading an OS kernel into memory, often along with support files such as a Linux initial RAM disk (initrd) file, and starting the kernel running.

On the other hand, a boot manager presents a menu of options to boot into, along with some other way
to control the boot process. After the user selects an option, the boot manager will hand off control
of booting to the OS.