features: Optimized Repositories
This fully deprecates the "CachyOS Repositories" section. In this page, I have merged "What are CachyOS
Repositories" and "How to Add CachyOS Repositories" into one page. To not bloat the page, I have decided
to remove the whole manual installation instructions and instead link to it under "Learn more".

P.S. for Aarrayy:
I moved the wiki link to x86-64vX instructions sets to "Learn More" and removed the tip about Skylake because
it is actually explained in Skylake's wiki page that only a few variants of Skylake have AVX512