wiki: Add new section documenting CachyOS's features (#58)
* wiki: Add new section documenting CachyOS's features

This initial commit moves the kernel and modules page to a new section labeled "CachyOS Features".
sched-ext tutorial page was moved to Configuration and the whole Kernel section was removed as it was
rendered obsolete by this commit.

Co-authored-by: Aarrayy <aarrayyzuelo@proton.me>

* features: Add CachyOS Settings

This section obsoletes some entries in configuration/general_system_tweaks, namely enabling KSM and core dumps.
For the latter, I have added comments in the respected file and made a general guide on how to modify sysctl
values

This section also complete obsoletes cachyos_repositories/other_optimizations as it effectively replaces the section.

Co-authored-by: Aarrayy <aarrayyzuelo@proton.me>

* features: Optimized Repositories

This fully deprecates the "CachyOS Repositories" section. In this page, I have merged "What are CachyOS
Repositories" and "How to Add CachyOS Repositories" into one page. To not bloat the page, I have decided
to remove the whole manual installation instructions and instead link to it under "Learn more".

P.S. for Aarrayy:
I moved the wiki link to x86-64vX instructions sets to "Learn More" and removed the tip about Skylake because
it is actually explained in Skylake's wiki page that only a few variants of Skylake have AVX512

* features: cachy-chroot

---------

Co-authored-by: Aarrayy <aarrayyzuelo@proton.me>