filesystem: Remove note
This note is somewhat misleading. While it is true that the performance between various filesystems are
unseen (for the most part) on NVMEs, the opposite can be true on other storage devices.

Therefore to avoid confusion and without going to too much detail, I will opt to remove the note.