features/chwd: Auto configure without explicitly passing a class ID
chwd supports auto configuring without needing to explicitly pass a class ID.
Due to this and the ability to install a specific profile, it seems a bit redundant to document this as well,
as some might categorize it as overinformation. The removed passage might be added back in the future, but for now
the current information provided is sufficient.

Signed-off-by: Eric Naim <dnaim@cachyos.org>