installation/installation_on_root: Add dual boot setup
On systemd-boot and GRUB, there needs to be additional setup before both boot managers to detect Windows and add it as a
valid boot entry. For systemd-boot, manually copying over Windows's EFI folders to `/boot/EFI` is necessary while we need
to install and enable `os-prober` with GRUB. As far as I'm aware, rEFInd has full auto-detection for this so no additional
setup is necessary.

I've also removed `installation_dualboot` because that page only has a link to a youtube video on it, and moved that video
link to `installation_on_root` instead.

Signed-off-by: Eric Naim <dnaim@cachyos.org>