installation/installation_on_root: Add dual boot setup (#71)
* installation: Remove virtual machine installation guide

As far as I can see, these guides only show how to setup a virtual machine without actually having a specific installation
guide. This is a bit out of place and I don't see much of a use for these two pages, the main installation page should cover
all usecases.

Signed-off-by: Eric Naim <dnaim@cachyos.org>

* chore: Remove outdated page in translations folder

The main page was previously removed and made redundant with #42, but I forgot to remove the corresponding translation of it.

Signed-off-by: Eric Naim <dnaim@cachyos.org>

* installation/installation_on_root: Add dual boot setup

On systemd-boot and GRUB, there needs to be additional setup before both boot managers to detect Windows and add it as a
valid boot entry. For systemd-boot, manually copying over Windows's EFI folders to `/boot/EFI` is necessary while we need
to install and enable `os-prober` with GRUB. As far as I'm aware, rEFInd has full auto-detection for this so no additional
setup is necessary.

I've also removed `installation_dualboot` because that page only has a link to a youtube video on it, and moved that video
link to `installation_on_root` instead.

Signed-off-by: Eric Naim <dnaim@cachyos.org>

---------

Signed-off-by: Eric Naim <dnaim@cachyos.org>