kernel: Rewrite Features
Mentioning patches is **NOT** a good way to advertise what features we offer. Firstly, we use short handles
for the patch names. These don't usually tell the whole story and can sometimes be misleading to what each patch does
(the names are still used because it's easy to understand internally and is easy to type out). Secondly, The descriptions
to describe what each patch does to compensate for the potentially misleading and very ambigous shorthandles are very long,
not to mention unsightly to read. Lastly, what patches we offer always change on the next major kernel version. This
means that to keep information accurate, we would need to update this section everytime. This is very bad for maintenance,
especially for a lazy bum like me, and futureproofing is practically non-existant.

The very tidy shortlist of features like this is much more elegant, especially to the user reading. I've listed most
of the notable features that we provide, most of them are also historically always included in the CachyOS patchset so
it doesn't hurt even if someday it's not included anymore.

Signed-off-by: Eric Naim <dnaim@cachyos.org>